Beef up precision in computePixelStats (#3353)
We should have been doing this particular multiplication as doubles, and
in just a couple spots, it slightly changes the answer.

+++ src/libOpenImageIO/imagebufalgo_compare.cpp
@@ -89,7 +89,7 @@ val(ImageBufAlgo::PixelStats& p, int c, float value)
     }
     ++p.finitecount[c];
     p.sum[c] += value;
-    p.sum2[c] += value * value;
+    p.sum2[c] += double(value) * double(value);
     p.min[c] = std::min(value, p.min[c]);
     p.max[c] = std::max(value, p.max[c]);
 }
